Refactorv2
