+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import ProductList from './components/ProductList';
+import ProductList from './components/ProductList/ProductList';
 
 function App() {
   return (
