+++ src/components/ProductList/viewmodel.ts
